Rename file
Printing out terms of the Fibonacci sequence seems to specific of an
activity to be a class onto itself—it seems like it should just be a
method as part of a larger class. I mean you could have a
FibonacciSequence class, which then contains a FibonacciSequence
method, but that seems somehow redundant, and not really how classes
are supposed to be designed. Better to have a more general Sequences
class (which might contain methods that generate any number of
sequences in addition to the Fibonacci sequence).